pyproject.toml
[tool.poetry]
authors = ["Jangwon Seo <nanpuhaha@gmail.com>"]
description = "Unofficial Python API client for MapleStory"
name = "maplestory-sdk-py"
packages = [{ include = "maplestory", from = "." }]
readme = "README.md"
version = "0.1.1"

[tool.poetry.dependencies]
python = "^3.12"
bs4 = "^0.0.2"
httpx = "^0.26.0"
lxml = "^5.1.0"
pillow = "^10.2.0"
pydantic = "^2.5.3"
pydantic-settings = "^2.1.0"
tenacity = "^8.2.3"

[tool.poetry.group.dev.dependencies]
autoflake = "^2.2.1"
black = "^23.12.1"
devtools = "^0.12.2"
furo = "^2023.9.10"
ipykernel = "^6.29.0"
isort = "^5.13.2"
pytest = "^7.4.4"
pytest-cov = "^4.1.0"
pytest-env = "^1.1.3"
pytest-sugar = "^0.9.7"
rich = "^13.7.0"

[tool.black]
color = true
line-length = 88
target-version = ['py312']

[tool.isort]
atomic = true
color_output = true
float_to_top = true
line_length = 88
profile = "black"
py_version = 312
skip_gitignore = true
src_paths = ["maplestory", "tests"]

[tool.pytest.ini_options]
addopts = "-ra -s"
minversion = "6.0"
pythonpath = "."
testpaths = ["tests"]

[build-system]
build-backend = "poetry.core.masonry.api"
requires = ["poetry-core"]
